[refactor](pipeline) Delete non-pipeline data sink
