+++ be/src/pipeline/exec/exchange_sink_buffer.h
@@ -51,7 +51,6 @@ class ExchangeSinkLocalState;
 } // namespace pipeline
 
 namespace vectorized {
-class VDataStreamSender;
 class PipChannel;
 
 template <typename T>
