+++ be/src/pipeline/exec/group_commit_block_sink_operator.h
@@ -18,7 +18,7 @@
 #pragma once
 
 #include "operator.h"
-#include "vec/sink/group_commit_block_sink.h"
+#include "runtime/group_commit_mgr.h"
 
 namespace doris::pipeline {
