+++ be/src/runtime/buffer_control_block.cpp
@@ -103,134 +103,149 @@ Status BufferControlBlock::init() {
     return Status::OK();
 }
 
-bool BufferControlBlock::can_sink() {
-    std::unique_lock<std::mutex> l(_lock);
-    return _get_batch_queue_empty() || _buffer_rows < _buffer_limit || _is_cancelled;
-}
-
 Status BufferControlBlock::add_batch(std::unique_ptr<TFetchDataResult>& result) {
-    std::unique_lock<std::mutex> l(_lock);
+    {
+        std::unique_lock<std::mutex> l(_lock);
 
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
-    }
-
-    int num_rows = result->result_batch.rows.size();
-
-    while ((!_fe_result_batch_queue.empty() && _buffer_rows > _buffer_limit) && !_is_cancelled) {
-        _data_removal.wait_for(l, std::chrono::seconds(1));
-    }
-
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
-    }
-
-    if (_waiting_rpc.empty()) {
-        // Merge result into batch to reduce rpc times
-        if (!_fe_result_batch_queue.empty() &&
-            ((_fe_result_batch_queue.back()->result_batch.rows.size() + num_rows) <
-             _buffer_limit) &&
-            !result->eos) {
-            std::vector<std::string>& back_rows = _fe_result_batch_queue.back()->result_batch.rows;
-            std::vector<std::string>& result_rows = result->result_batch.rows;
-            back_rows.insert(back_rows.end(), std::make_move_iterator(result_rows.begin()),
-                             std::make_move_iterator(result_rows.end()));
-        } else {
-            _fe_result_batch_queue.push_back(std::move(result));
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
         }
-        _buffer_rows += num_rows;
-    } else {
-        auto ctx = _waiting_rpc.front();
-        _waiting_rpc.pop_front();
-        ctx->on_data(result, _packet_num);
-        _packet_num++;
+
+        int num_rows = result->result_batch.rows.size();
+
+        while ((!_fe_result_batch_queue.empty() && _buffer_rows > _buffer_limit) &&
+               !_is_cancelled) {
+            _data_removal.wait_for(l, std::chrono::seconds(1));
+        }
+
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
+        }
+
+        if (_waiting_rpc.empty()) {
+            // Merge result into batch to reduce rpc times
+            if (!_fe_result_batch_queue.empty() &&
+                ((_fe_result_batch_queue.back()->result_batch.rows.size() + num_rows) <
+                 _buffer_limit) &&
+                !result->eos) {
+                std::vector<std::string>& back_rows =
+                        _fe_result_batch_queue.back()->result_batch.rows;
+                std::vector<std::string>& result_rows = result->result_batch.rows;
+                back_rows.insert(back_rows.end(), std::make_move_iterator(result_rows.begin()),
+                                 std::make_move_iterator(result_rows.end()));
+            } else {
+                _fe_result_batch_queue.push_back(std::move(result));
+            }
+            _buffer_rows += num_rows;
+        } else {
+            auto ctx = _waiting_rpc.front();
+            _waiting_rpc.pop_front();
+            ctx->on_data(result, _packet_num);
+            _packet_num++;
+        }
     }
+    _update_dependency();
     return Status::OK();
 }
 
 Status BufferControlBlock::add_arrow_batch(std::shared_ptr<arrow::RecordBatch>& result) {
-    std::unique_lock<std::mutex> l(_lock);
+    {
+        std::unique_lock<std::mutex> l(_lock);
 
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
+        }
+
+        int num_rows = result->num_rows();
+
+        while ((!_arrow_flight_batch_queue.empty() && _buffer_rows > _buffer_limit) &&
+               !_is_cancelled) {
+            _data_removal.wait_for(l, std::chrono::seconds(1));
+        }
+
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
+        }
+
+        // TODO: merge RocordBatch, ToStructArray -> Make again
+
+        _arrow_flight_batch_queue.push_back(std::move(result));
+        _buffer_rows += num_rows;
+        _data_arrival.notify_one();
     }
-
-    int num_rows = result->num_rows();
-
-    while ((!_arrow_flight_batch_queue.empty() && _buffer_rows > _buffer_limit) && !_is_cancelled) {
-        _data_removal.wait_for(l, std::chrono::seconds(1));
-    }
-
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
-    }
-
-    // TODO: merge RocordBatch, ToStructArray -> Make again
-
-    _arrow_flight_batch_queue.push_back(std::move(result));
-    _buffer_rows += num_rows;
-    _data_arrival.notify_one();
+    _update_dependency();
     return Status::OK();
 }
 
 void BufferControlBlock::get_batch(GetResultBatchCtx* ctx) {
-    std::lock_guard<std::mutex> l(_lock);
-    if (!_status.ok()) {
-        ctx->on_failure(_status);
-        return;
-    }
-    if (_is_cancelled) {
-        ctx->on_failure(Status::Cancelled("Cancelled"));
-        return;
-    }
-    if (!_fe_result_batch_queue.empty()) {
-        // get result
-        std::unique_ptr<TFetchDataResult> result = std::move(_fe_result_batch_queue.front());
-        _fe_result_batch_queue.pop_front();
-        _buffer_rows -= result->result_batch.rows.size();
-        _data_removal.notify_one();
+    {
+        std::lock_guard<std::mutex> l(_lock);
+        if (!_status.ok()) {
+            ctx->on_failure(_status);
+            _update_dependency();
+            return;
+        }
+        if (_is_cancelled) {
+            ctx->on_failure(Status::Cancelled("Cancelled"));
+            _update_dependency();
+            return;
+        }
+        if (!_fe_result_batch_queue.empty()) {
+            // get result
+            std::unique_ptr<TFetchDataResult> result = std::move(_fe_result_batch_queue.front());
+            _fe_result_batch_queue.pop_front();
+            _buffer_rows -= result->result_batch.rows.size();
+            _data_removal.notify_one();
 
-        ctx->on_data(result, _packet_num);
-        _packet_num++;
-        return;
+            ctx->on_data(result, _packet_num);
+            _packet_num++;
+            _update_dependency();
+            return;
+        }
+        if (_is_close) {
+            ctx->on_close(_packet_num, _query_statistics.get());
+            _update_dependency();
+            return;
+        }
+        // no ready data, push ctx to waiting list
+        _waiting_rpc.push_back(ctx);
     }
-    if (_is_close) {
-        ctx->on_close(_packet_num, _query_statistics.get());
-        return;
-    }
-    // no ready data, push ctx to waiting list
-    _waiting_rpc.push_back(ctx);
+    _update_dependency();
 }
 
 Status BufferControlBlock::get_arrow_batch(std::shared_ptr<arrow::RecordBatch>* result) {
-    std::unique_lock<std::mutex> l(_lock);
-    if (!_status.ok()) {
-        return _status;
-    }
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
-    }
+    {
+        std::unique_lock<std::mutex> l(_lock);
+        if (!_status.ok()) {
+            return _status;
+        }
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
+        }
 
-    while (_arrow_flight_batch_queue.empty() && !_is_cancelled && !_is_close) {
-        _data_arrival.wait_for(l, std::chrono::seconds(1));
-    }
+        while (_arrow_flight_batch_queue.empty() && !_is_cancelled && !_is_close) {
+            _data_arrival.wait_for(l, std::chrono::seconds(1));
+        }
 
-    if (_is_cancelled) {
-        return Status::Cancelled("Cancelled");
-    }
+        if (_is_cancelled) {
+            return Status::Cancelled("Cancelled");
+        }
 
-    if (!_arrow_flight_batch_queue.empty()) {
-        *result = std::move(_arrow_flight_batch_queue.front());
-        _arrow_flight_batch_queue.pop_front();
-        _buffer_rows -= (*result)->num_rows();
-        _data_removal.notify_one();
-        _packet_num++;
-        return Status::OK();
-    }
+        if (!_arrow_flight_batch_queue.empty()) {
+            *result = std::move(_arrow_flight_batch_queue.front());
+            _arrow_flight_batch_queue.pop_front();
+            _buffer_rows -= (*result)->num_rows();
+            _data_removal.notify_one();
+            _packet_num++;
+            _update_dependency();
+            return Status::OK();
+        }
 
-    // normal path end
-    if (_is_close) {
-        return Status::OK();
+        // normal path end
+        if (_is_close) {
+            _update_dependency();
+            return Status::OK();
+        }
     }
     return Status::InternalError("Abnormal Ending");
 }
@@ -266,42 +281,15 @@ void BufferControlBlock::cancel() {
         ctx->on_failure(Status::Cancelled("Cancelled"));
     }
     _waiting_rpc.clear();
-}
-
-Status PipBufferControlBlock::add_batch(std::unique_ptr<TFetchDataResult>& result) {
-    RETURN_IF_ERROR(BufferControlBlock::add_batch(result));
-    _update_dependency();
-    return Status::OK();
-}
-
-Status PipBufferControlBlock::add_arrow_batch(std::shared_ptr<arrow::RecordBatch>& result) {
-    RETURN_IF_ERROR(BufferControlBlock::add_arrow_batch(result));
-    _update_dependency();
-    return Status::OK();
-}
-
-void PipBufferControlBlock::get_batch(GetResultBatchCtx* ctx) {
-    BufferControlBlock::get_batch(ctx);
     _update_dependency();
 }
 
-Status PipBufferControlBlock::get_arrow_batch(std::shared_ptr<arrow::RecordBatch>* result) {
-    RETURN_IF_ERROR(BufferControlBlock::get_arrow_batch(result));
-    _update_dependency();
-    return Status::OK();
-}
-
-void PipBufferControlBlock::cancel() {
-    BufferControlBlock::cancel();
-    _update_dependency();
-}
-
-void PipBufferControlBlock::set_dependency(
+void BufferControlBlock::set_dependency(
         std::shared_ptr<pipeline::Dependency> result_sink_dependency) {
     _result_sink_dependency = result_sink_dependency;
 }
 
-void PipBufferControlBlock::_update_dependency() {
+void BufferControlBlock::_update_dependency() {
     if (_result_sink_dependency &&
         (_batch_queue_empty || _buffer_rows < _buffer_limit || _is_cancelled)) {
         _result_sink_dependency->set_ready();
@@ -309,11 +297,6 @@ void PipBufferControlBlock::_update_dependency() {
                (!_batch_queue_empty && _buffer_rows < _buffer_limit && !_is_cancelled)) {
         _result_sink_dependency->block();
     }
-}
-
-void PipBufferControlBlock::_update_batch_queue_empty() {
-    _batch_queue_empty = _fe_result_batch_queue.empty() && _arrow_flight_batch_queue.empty();
-    _update_dependency();
 }
 
 } // namespace doris
