+++ be/src/runtime/result_buffer_mgr.cpp
@@ -68,7 +68,7 @@ Status ResultBufferMgr::init() {
 
 Status ResultBufferMgr::create_sender(const TUniqueId& query_id, int buffer_size,
                                       std::shared_ptr<BufferControlBlock>* sender,
-                                      bool enable_pipeline, int exec_timout) {
+                                      int exec_timout) {
     *sender = find_control_block(query_id);
     if (*sender != nullptr) {
         LOG(WARNING) << "already have buffer control block for this instance " << query_id;
@@ -77,11 +77,7 @@ Status ResultBufferMgr::create_sender(const TUniqueId& query_id, int buffer_size
 
     std::shared_ptr<BufferControlBlock> control_block = nullptr;
 
-    if (enable_pipeline) {
-        control_block = std::make_shared<PipBufferControlBlock>(query_id, buffer_size);
-    } else {
-        control_block = std::make_shared<BufferControlBlock>(query_id, buffer_size);
-    }
+    control_block = std::make_shared<BufferControlBlock>(query_id, buffer_size);
 
     {
         std::unique_lock<std::shared_mutex> wlock(_buffer_map_lock);
