+++ be/src/vec/sink/load_stream_map_pool.h
@@ -48,7 +48,6 @@
 
 #include "common/config.h"
 #include "common/status.h"
-#include "exec/data_sink.h"
 #include "exec/tablet_info.h"
 #include "gutil/ref_counted.h"
 #include "runtime/exec_env.h"
