+++ be/src/vec/sink/load_stream_stub.h
@@ -52,7 +52,6 @@
 
 #include "common/config.h"
 #include "common/status.h"
-#include "exec/data_sink.h"
 #include "exec/tablet_info.h"
 #include "gutil/ref_counted.h"
 #include "runtime/exec_env.h"
