+++ be/src/vec/sink/varrow_flight_result_writer.cpp
@@ -94,10 +94,6 @@ Status VArrowFlightResultWriter::write(Block& input_block) {
     return status;
 }
 
-bool VArrowFlightResultWriter::can_sink() {
-    return _sinker->can_sink();
-}
-
 Status VArrowFlightResultWriter::close(Status st) {
     COUNTER_SET(_sent_rows_counter, _written_rows);
     COUNTER_UPDATE(_bytes_sent_counter, _bytes_sent);
