+++ be/src/vec/sink/vmysql_result_writer.cpp
@@ -212,11 +212,6 @@ Status VMysqlResultWriter<is_binary_format>::write(Block& input_block) {
 }
 
 template <bool is_binary_format>
-bool VMysqlResultWriter<is_binary_format>::can_sink() {
-    return _sinker->can_sink();
-}
-
-template <bool is_binary_format>
 Status VMysqlResultWriter<is_binary_format>::close(Status) {
     COUNTER_SET(_sent_rows_counter, _written_rows);
     COUNTER_UPDATE(_bytes_sent_counter, _bytes_sent);
